Added makefile and changed float to double
The reading function needs to read the input as double to obtain the required precision.

+++ Code/auxiliar_func.h
@@ -3,10 +3,13 @@
 
 #include <stdio.h>
 #include <vector>
+#include <tuple>
 #include <math.h>
 
 #include <fstream>
 #include <string>
+#include <stdlib.h> 
+
 
 
 
@@ -21,6 +24,6 @@ vector<vector<int>> compute_dist_matrix(vector<double> coord_x, vector<double> c
 
 vector<vector<int>> compute_dist_matrix(vector<float> coord_x, vector<float> coord_y);
 
-vector<vector<int>> compute_dist_matrix(vector<tuple<int, float, float>> instance);
+vector<vector<int>> compute_dist_matrix(vector<tuple<int, double, double>> instance);
 
 #endif
